style(mdx): update integration components
